+++ app/components/helpMenu.js
@@ -30,21 +30,21 @@
                                   ' ng-show="vm.show(item)">'+
                               '<div class="filters" help-inverter>'+
                                   '<icon svg="item.svg" color="item.color||obj.iconcolor.hex"></icon>'+
-                                  "{{item.langButton?' '+app.config.lang.buttons[item.langButton]:''}}"+
+                                  "{{item.langButton?' '+app.config.lang.button[item.langButton]:''}}"+
                               '</div>'+
                               '<div ng-class="app.styler.helpMenuOptionClasses"'+
                                     ' ng-style="app.styler.helpTextColorBorder"'+
                                     ' ng-show="hover && item.menu"'+
                                     ' ng-click="hover=false">'+
                                  '<a ng-repeat="option in item.menu" help-inverter'+
-                                    ' ng-click="option.click()" >{{option.text}}</a>'+
+                                    ' ng-click="option.click()" >{{app.config.lang[vm.view][option.text]}}</a>'+
                               '</div>'+
                           '</span>'+
                           '<ng-transclude></ng-transclude>'+
                       '</div>',
             controller: controller,
             controllerAs: 'vm',
-            bindToController: { list: '=', obj: '=' },
+            bindToController: { list: '=', obj: '=', view: '@' },
         };
         return directive;
     }