+++ app/components/themeImage.js
@@ -198,7 +198,7 @@
             {
                 if (config.edit)
                 {
-                    vm.title = 'Click to turn '+config.lang.md_labels[vm.md]+' OFF';
+                    vm.title = 'Click to turn '+config.lang.md_label[vm.md]+' OFF';
                 }
                 if (vm.obj.div)
                 {
@@ -213,7 +213,7 @@
             {
                 if (config.edit)
                 {
-                    vm.title = 'Click to turn '+config.lang.md_labels[vm.md]+' ON';
+                    vm.title = 'Click to turn '+config.lang.md_label[vm.md]+' ON';
                 }
 
                 if (vm.obj.div)